refactor: divide docs and added attributes
